requirements.txt

nidaqmx~=0.5.8
numpy~=1.22.1
pyserial~=3.4
matplotlib~=3.2.1